:bug: Fix exporter crashing during json parsing

+++ main.go
@@ -48,10 +48,10 @@ type videoBridgeStats struct {
 	OctoVersion                           int     `json:"octo_version"`
 	OctoConferences                       int     `json:"octo_conferences"`
 	OctoEndpoints                         int     `json:"octo_endpoints"`
-	OctoReceiveBitrate                    int     `json:"octo_receive_bitrate"`
-	OctoReceivePacketRate                 int     `json:"octo_receive_packet_rate"`
-	OctoSendBitrate                       int     `json:"octo_send_bitrate"`
-	OctoSendPacketRate                    int     `json:"octo_send_packet_rate"`
+	OctoReceiveBitrate                    float64 `json:"octo_receive_bitrate"`
+	OctoReceivePacketRate                 float64 `json:"octo_receive_packet_rate"`
+	OctoSendBitrate                       float64 `json:"octo_send_bitrate"`
+	OctoSendPacketRate                    float64 `json:"octo_send_packet_rate"`
 	TotalBytesReceivedOcto                int     `json:"total_bytes_received_octo"`
 	TotalBytesSentOcto                    int     `json:"total_bytes_sent_octo"`
 	TotalPacketsDroppedOcto               int     `json:"total_packets_dropped_octo"`
